+++ src/main/java/the/husky/web/servlet/vehicleservlet/DeleteVehicleServlet.java
@@ -4,10 +4,9 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import the.husky.entity.user.User;
 import the.husky.entity.vehicle.Vehicle;
-import the.husky.service.UserService;
 import the.husky.service.VehicleService;
+import the.husky.web.auth.UserAuthenticate;
 
 import java.io.IOException;
 
@@ -19,29 +18,18 @@ public class DeleteVehicleServlet extends HttpServlet {
     }
 
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String idParam = request.getParameter("vehicle_id");
-        int id;
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        if (UserAuthenticate.isAuthenticate(request)) {
+            int id = Integer.parseInt(request.getParameter("vehicle_id"));
 
-        try {
-            id = Integer.parseInt(idParam);
-        } catch (NumberFormatException e) {
-            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid 'id' parameter");
-            return;
+            Vehicle vehicle = service.getById(id);
+            if (vehicle != null) {
+                service.delete(id);
+                response.sendRedirect("/vehicle/all");
+            } else {
+                response.sendError(HttpServletResponse.SC_NOT_FOUND, "Vehicle not found");
+            }
         }
-
-        Vehicle vehicle = service.getById(id);
-
-        if (vehicle != null) {
-            service.delete(id);
-            response.sendRedirect("/vehicle/all");
-        } else {
-            response.sendError(HttpServletResponse.SC_NOT_FOUND, "Vehicle not found");
-        }
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        doGet(request, response);
+        response.sendRedirect("/login");
     }
 }
